+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Prefix your message with one of the following:
 
 - [Added] Microsoft Outlook detection.
 - [Added] Hey detection.
+- [Added] Yandex detection.
 
 ## v0.0.1 - 2022-01-15